backend/pyproject.toml
[project]
name = "nonware-backend"
version = "0.1.0"
description = "Django backend for Nonware portfolio"
readme = "README.md"
requires-python = ">=3.14.3"
dependencies = [
    "django>=6.0,<6.1",
    "djangorestframework>=3.16,<3.17",
    "django-cors-headers>=4.7,<4.8",
    "psycopg[binary]>=3.2,<3.3",
    "gunicorn>=23.0,<24.0",
    "whitenoise>=6.9,<7.0",
    "python-dotenv>=1.1,<1.2",
    "django-ninja>=1.6.1",
    "django-storages>=1.14.6",
    "google-cloud-storage>=3.10.0",
]

[dependency-groups]
dev = [
    "ruff>=0.11,<0.12",
    "pytest>=8.3,<8.4",
    "pytest-django>=4.11,<4.12",
]
